adaptive-onboarding/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
sentence-transformers
scikit-learn
numpy
networkx
pdfplumber
python-docx
groq
python-dotenv
httpx
